Split actions into separate action creators
